requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
psutil==5.9.6
pyg-lib==0.3.0+pt112cu116
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==1.12.0+cu116
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torch_geometric==2.4.0
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zipp==3.17.0
